Change np from int to size_t.

+++ src/boundary/boundary_p.cc
@@ -67,9 +67,9 @@ boundary_p_kokkos(
   // Temporary store for local particle injectors
   // FIXME: Ugly static usage
   static particle_injector_t * RESTRICT ALIGNED(16) ci = NULL;
-  static int max_ci = 0;
+  static size_t max_ci = 0;
 
-  int n_send[6], n_recv[6], n_ci;
+  size_t n_send[6], n_recv[6], n_ci;
 
   species_t * sp;
   int face;
@@ -158,7 +158,7 @@ boundary_p_kokkos(
     // more flexible though in the future (especially given above the
     // above overalloc).
 
-    int nm = 0; LIST_FOR_EACH( sp, sp_list ) nm += sp->nm;
+    size_t nm = 0; LIST_FOR_EACH( sp, sp_list ) nm += sp->nm;
 
     for( face=0; face<6; face++ )
       if( shared[face] ) {
@@ -205,7 +205,7 @@ boundary_p_kokkos(
         for( ; nm; pm--, nm-- )
         {
             //int i = pm->i;
-            int copy_index = nm -1;
+            size_t copy_index = nm -1;
 
             //int voxel = p0[i].i;
             int voxel = particle_send(copy_index);
@@ -390,7 +390,7 @@ boundary_p_kokkos(
     particle_mover_t * RESTRICT ALIGNED(32) sp_pm[MAX_SP];
     //float sp_q[MAX_SP];
     //int sp_np[MAX_SP];
-    int sp_nm[MAX_SP];
+    size_t sp_nm[MAX_SP];
 
     if( num_species( sp_list ) > MAX_SP )
     {
@@ -412,7 +412,7 @@ boundary_p_kokkos(
       //particle_t          * RESTRICT ALIGNED(32) p;
       particle_mover_t    * RESTRICT ALIGNED(16) pm;
       const particle_injector_t * RESTRICT ALIGNED(16) pi;
-      int nm, n, id;
+      size_t nm, n, id;
 
       face++; if( face==7 ) face = 0;
       if( face==6 ) pi = ci, n = n_ci;
@@ -439,7 +439,7 @@ boundary_p_kokkos(
         auto& particle_send = sp_[id]->k_pc_h;
         auto& particle_send_i = sp_[id]->k_pc_i_h;
 
-        int write_index = sp_[id]->num_to_copy;
+        size_t write_index = sp_[id]->num_to_copy;
 
         // Write out received particle data
         //p[np].dx=pi->dx;
@@ -491,7 +491,7 @@ boundary_p_kokkos(
                 sp_[id]->q
         );
 
-        int keep_id = nm + ret_code - 1;
+        size_t keep_id = nm + ret_code - 1;
         sp_nm[id] = keep_id+1; // +1 to convert from index to count:w
 
         if (ret_code)
